# storage/api-docs/hrm-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini ERP HRM API",
        "description": "Human Resources Management API Documentation for Mini ERP App - Includes Employee Management, Attendance, Leave, Payroll, Reimbursements, Resignations, Office Locations, and Analytics & Reporting",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/platform/hrm/designations": {
            "get": {
                "tags": [
                    "HRM Designations"
                ],
                "summary": "List all designations",
                "operationId": "59b36e47a66a9a1e09f2aa36861b98ac",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HRM Designations"
                ],
                "summary": "Create a new designation",
                "operationId": "86230d9e649e9fdabd6b0dd4756007a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Designation name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Optional description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Designation created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/platform/hrm/designations/{id}": {
            "get": {
                "tags": [
                    "HRM Designations"
                ],
                "summary": "Get designation details",
                "operationId": "7129a5aefa011ef7af1b15d832dad628",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Designation not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "HRM Designations"
                ],
                "summary": "Update a designation",
                "operationId": "bb09b9ff7abccb3d45d7f5ff44b9e4b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Designation name",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Optional description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Designation updated"
                    },
                    "404": {
                        "description": "Designation not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "HRM Designations"
                ],
                "summary": "Delete a designation",
                "operationId": "2d3486a64f443d9a9c5d0dbdf7e41858",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Designation deleted"
                    },
                    "404": {
                        "description": "Designation not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use a token from Sanctum auth",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}